# spring-boot-demo/src/docs/asciidoc/swagger.json
{"swagger":"2.0","info":{"description":"Rest API Doc for Alarm Management","version":"1.0.0","title":"Alarm Management Rest API","termsOfService":"http://www.example.com/","contact":{"name":"Admin","url":"http://www.example.com/","email":"admin@example.com"}},"host":"localhost","basePath":"/","tags":[{"name":"Alarm","description":"Alarm Controller"},{"name":"AlarmHistory","description":"Alarm History Controller"}],"paths":{"/api/v2/alarms":{"get":{"tags":["Alarm"],"summary":"Query all alarms","description":"Query all alarms both active and inactive","operationId":"queryAlarmsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"oauth2schema":["query","trust"]}],"deprecated":false},"post":{"tags":["Alarm"],"summary":"Create an alarm","description":"Create an alarm by input body","operationId":"createAlarmUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"alarm","description":"alarm","required":true,"schema":{"$ref":"#/definitions/Alarm"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"201":{"description":"Created","schema":{"$ref":"#/definitions/Response"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"oauth2schema":["create","trust"]}],"deprecated":false},"put":{"tags":["Alarm"],"summary":"Update an alarm","description":"Update an alarm by input body","operationId":"updateAlarmUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"alarm","description":"alarm","required":true,"schema":{"$ref":"#/definitions/Alarm"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"oauth2schema":["update","trust"]}],"deprecated":false},"delete":{"tags":["Alarm"],"summary":"Delete alarms conditionally","description":"Delete alarms by status, interval and timestamp","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"interval","in":"query","description":"interval","required":false,"type":"integer","format":"int32"},{"name":"timeUnit","in":"query","description":"timeUnit","required":false,"type":"string","enum":["NANOSECONDS","MICROSECONDS","MILLISECONDS","SECONDS","MINUTES","HOURS","DAYS"]}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Response"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2schema":["delete","trust"]}],"deprecated":false}},"/api/v2/alarms/history":{"delete":{"tags":["AlarmHistory"],"summary":"Delete alarm operation history","description":"Delete alarm operation history conditionally","operationId":"deleteAlarmHistoryUsingDELETE","produces":["*/*"],"parameters":[{"name":"module","in":"query","description":"module","required":false,"type":"string"},{"name":"uptoTimestamp","in":"query","description":"uptoTimestamp","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaseResult"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"oauth2schema":["delete","trust"]}],"deprecated":false}},"/api/v2/alarms/history/all":{"get":{"tags":["AlarmHistory"],"summary":"Query all alarm operation history","description":"Query all alarm operation history conditionally","operationId":"queryAllAlarmHistoryUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/BaseResult"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"oauth2schema":["query","trust"]}],"deprecated":false}}},"securityDefinitions":{"oauth2schema":{"type":"oauth2","tokenUrl":"http://127.0.0.1:8080/tni-api/oauth/token","flow":"password","scopes":{"create":"create all","query":"query all","update":"update all","delete":"delete all","trust":"trust all"}}},"definitions":{"Alarm":{"type":"object","properties":{"auid":{"type":"string"},"id":{"type":"integer","format":"int64"},"instanceId":{"type":"integer","format":"int32"},"json":{"type":"string"},"module":{"type":"string"},"name":{"type":"string"},"severity":{"type":"string","enum":["INFO","FATAL","WARNING"]},"status":{"type":"string","enum":["RAISED","CLEARED","MANUAL_CLEARED"]},"ts":{"type":"string"}},"title":"Alarm","description":"Alarm entity"},"BaseResult":{"type":"object","required":["code","msg"],"properties":{"code":{"type":"integer","format":"int32","example":0,"description":"Response code"},"data":{"type":"object","description":"Response data"},"msg":{"type":"string","example":"succeed","description":"Response message"}},"title":"BaseResult","description":"Response Entity"},"Response":{"type":"object","properties":{"entity":{"type":"object"},"mediaType":{"type":"string"},"status":{"type":"integer","format":"int32"}},"title":"Response","description":"Response wrapper model"}}}